+++ src/Controller/Admin/PagesCategoriesController.php
@@ -58,7 +58,7 @@ class PagesCategoriesController extends AppController
     public function isAuthorized($user = null)
     {
         //Only admins can delete pages categories
-        if ($this->request->isAction('delete')) {
+        if ($this->request->isDelete()) {
             return $this->Auth->isGroup('admin');
         }