+++ src/Controller/Admin/PhotosAlbumsController.php
@@ -40,7 +40,7 @@ class PhotosAlbumsController extends AppController
     public function isAuthorized($user = null)
     {
         //Only admins can delete albums
-        if ($this->request->isAction('delete')) {
+        if ($this->request->isDelete()) {
             return $this->Auth->isGroup('admin');
         }