+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # 2.x branch
 ## 2.13 branch
 ### 2.13.1
+* added `is('add')`, `is('delete')`, `is('edit')`, `is('index')` and
+    `is('view')` detectors;
 * the admin sidebar is cached only if debugging is disabled;
 * improved admin routes. They are automatically handled by CakePHP;
 * fixed other code for CakePHP Code Sniffer.